Refactore list creation components
