+++ lib/createlists.js
@@ -10,4 +10,15 @@ function createCategoryList(collection) {
     return filteredCategoryChoices;
 }
 
+function createPasswordList(documents) {
+    const choices = documents.map((document) => {
+        return { 
+            name: document.name,
+            value: document._id
+        };
+    });
+    return choices;
+}
+
+exports.createPasswordList = createPasswordList;
 exports.createCategoryList = createCategoryList;
